Merging updates from dynamics into main.
The dynamics package is included to allow for controlled changes
to be made to various parameters throughout the MadNet system.
These changes may be desired at some point in the future.

Some of the changes include allowing for fees in application
objects, including transaction fees.

There were also some changes to crypto/ to ensure the objects
behave as expected and do not panic under certain circumstances.